app/requirements.txt
pandas==2.2.2
Pillow==10.3.0
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
jupyter==1.0.0
flask==3.0.3
python-dotenv
scikit-learn
